moved methods to BaseActivity
