+++ app/src/main/java/hu/ait/macweekly/BaseActivity.java
@@ -1,26 +1,44 @@
 package hu.ait.macweekly;
 
+import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.ProgressDialog;
 import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.content.Intent;
 import android.content.IntentFilter;
+import android.content.pm.PackageManager;
+import android.content.pm.ResolveInfo;
+import android.content.res.Resources;
 import android.graphics.Color;
 import android.graphics.drawable.Drawable;
+import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
 import android.support.annotation.Nullable;
 import android.support.design.widget.Snackbar;
+import android.support.v4.content.FileProvider;
 import android.support.v4.content.res.ResourcesCompat;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.RecyclerView;
+import android.util.DisplayMetrics;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.util.List;
+
 import butterknife.BindView;
 
 public class BaseActivity extends AppCompatActivity {
+
+    private final int SEND_EMAIL_REQUEST = 0;
 
     private ProgressDialog mProgressDialog;
     private View rootLayout;
@@ -77,7 +95,12 @@ public class BaseActivity extends AppCompatActivity {
 
         final boolean alertVal;
 
-        builder.setMessage(content).setTitle(title);
+        builder.setMessage(content).setTitle(title).setPositiveButton(R.string.confirm_email_failure, new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialogInterface, int i) {
+                //okay
+            }
+        });
         AlertDialog dialog = builder.create();
 
         dialog.show();
@@ -131,5 +154,132 @@ public class BaseActivity extends AppCompatActivity {
     public boolean onSupportNavigateUp() {
         onBackPressed();
         return true;
+    }
+
+    /**
+     * Check that a mail client is present
+     */
+    public boolean isMailClientPresent() {
+        Intent intent = new Intent(Intent.ACTION_SEND);
+        intent.setType("text/html");
+        final PackageManager packageManager = getPackageManager();
+        List<ResolveInfo> list = packageManager.queryIntentActivities(intent, 0);
+
+        if(list.size() == 0)
+            return false;
+        else
+            return true;
+    }
+
+    /**
+     * Send email feedback
+     */
+    public void sendFeedback() {
+        final String[] recipients = {"themacweeklyapp@gmail.com"};
+        long currentTime = System.currentTimeMillis();
+        Resources res = getResources();
+        String emailSubject = String.format(res.getString(R.string.feedback_email_subject), currentTime);
+
+        Intent sendFeedbackIntent = new Intent(Intent.ACTION_SEND); //We have to use ACTION_SEND to include a single attachment - via Android dev docs
+        sendFeedbackIntent.setType("text/plain");
+        sendFeedbackIntent.putExtra(Intent.EXTRA_EMAIL, recipients);
+        sendFeedbackIntent.putExtra(Intent.EXTRA_SUBJECT, emailSubject);
+
+        String phoneDesc;
+
+        FileOutputStream fos = null;
+        File reportInfoFile = null;
+        try {
+            reportInfoFile = new File(getFilesDir(), "tempReportFile.txt");
+            fos = openFileOutput("tempReportFile.txt", Context.MODE_PRIVATE);
+            String writerString = getPhoneDetailsString();
+            fos.write(writerString.getBytes());
+
+        } catch (IOException e) {
+//            showSnackbar(e.getMessage(), Snackbar.LENGTH_LONG);
+        } finally {
+            if (fos != null){
+                try {
+                    fos.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+
+        Uri fileUri = FileProvider.getUriForFile(this,
+                "hu.ait.macweekly.fileprovider",
+                reportInfoFile);
+
+        if (fileUri != null) {
+            sendFeedbackIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            sendFeedbackIntent.addFlags(Intent.FLAG_GRANT_WRITE_URI_PERMISSION);
+            BaseActivity.this.setResult(Activity.RESULT_OK, sendFeedbackIntent);
+            sendFeedbackIntent.putExtra(Intent.EXTRA_STREAM, fileUri);
+
+            String defaultBody = "Please write your thoughts or describe your issue."; //Feel free to change this.
+
+            sendFeedbackIntent.putExtra(Intent.EXTRA_TEXT, defaultBody);
+
+            if (isMailClientPresent()) {
+                startActivityForResult(Intent.createChooser(sendFeedbackIntent, res.getString(R.string.email_app_chooser)), SEND_EMAIL_REQUEST);
+            } else {
+                showAlertDialogue(res.getString(R.string.error_title), res.getString(R.string.no_email_app));
+            }
+        }
+
+    }
+
+    /**
+     * Format and return a string containing details about
+     * the user's phone
+     * @return
+     */
+    public String getPhoneDetailsString() {
+
+        Resources res = getResources();
+        StringBuilder writer = new StringBuilder();
+
+        writer.append(String.format(res.getString(R.string.phone_manufacturer), Build.MANUFACTURER));
+        writer.append(System.getProperty("line.separator"));
+
+        writer.append(String.format(res.getString(R.string.phone_os), Build.VERSION.RELEASE));
+        writer.append(System.getProperty("line.separator"));
+
+        DisplayMetrics displayMetrics = new DisplayMetrics();
+        getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
+        writer.append(String.format(res.getString(R.string.phone_screen_resolution), displayMetrics.heightPixels, displayMetrics.widthPixels));
+        writer.append(System.getProperty("line.separator"));
+
+        writer.append(String.format(res.getString(R.string.phone_model), Build.MODEL));
+        writer.append(System.getProperty("line.separator"));
+
+        writer.append(String.format(res.getString(R.string.phone_hardware), Build.HARDWARE));
+        writer.append(System.getProperty("line.separator"));
+
+        writer.append(String.format(res.getString(R.string.phone_serial), Build.SERIAL));
+        writer.append(System.getProperty("line.separator"));
+
+        writer.append(String.format(res.getString(R.string.phone_tags), Build.TAGS));
+        writer.append(System.getProperty("line.separator"));
+
+        return writer.toString();
+
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        if (requestCode == SEND_EMAIL_REQUEST) {
+            if (resultCode == RESULT_OK) {
+                File tempData = new File(getFilesDir(), "tempReportFile.txt");
+                if (tempData.exists()) {
+                    tempData.delete();
+                }
+//                this.showSnackbar(getResources().getString(R.string.email_sent), Snackbar.LENGTH_SHORT); //calling showSnackbar at the end is causing a crash
+            } else {
+                showAlertDialogue(getResources().getString(R.string.error_title),
+                        getResources().getString(R.string.email_not_sent));
+            }
+        }
     }
 }
