+++ app/src/main/res/values/strings.xml
@@ -141,6 +141,7 @@
     <string name="no_email_app">No email application detected - click to dismiss.</string>
     <string name="email_not_sent">Email not sent - click to dismiss.</string>
     <string name="email_app_chooser">Send email using: </string>
+    <string name="confirm_email_failure">Okay</string>
 
 
     <!-- TODO: Remove or change this placeholder text -->
