+++ app/src/main/res/xml/filepaths.xml
@@ -1,0 +1,5 @@
+<paths xmlns:android="http://schemas.android.com/apk/res/android">
+    <files-path
+        name="mFiles"
+        path="." />
+</paths>
